Separate archive from database export

+++ backupTool.py
@@ -2,6 +2,7 @@ __author__ = 'marc'
 from constants import DATA_FOLDER
 import csv
 import db.datastore as db
+import errno
 import os
 import re
 import shutil
@@ -9,7 +10,7 @@ import sqlite3 as lite
 import tarfile
 import time
 
-TABLES = {
+_TABLES = {
 	'hashmap': ['hash', 'filepath'],
 	'profile': ['id', 'serializedUserInfo'],
 	'listings': ['id', 'serializedListings'],
@@ -26,8 +27,17 @@ TABLES = {
 }
 
 def _getDatabase():
+	"""Retrieves the OpenBazaar database file."""
 	Database = db.Database()
 	return Database.DATABASE
+
+def silentRemove(filename):
+	"""Silently removes a file if it exists."""
+	try:
+		os.remove(filename)
+	except OSError as e:
+		if e.errno != errno.ENOENT: # ENOENT: no such file or directory
+			raise
 
 def _exportDatabaseToCsv(tablesAndColumns):
 	"""Reads the database for all given tables and stores them as CSV files."""
@@ -48,34 +58,38 @@ def _exportDatabaseToCsv(tablesAndColumns):
 				writer.writerows(data)
 	return result
 
-def backup(tableList=None, output=None):
-	"""Archives given tables and files in a single tar archive."""
+def backupFiles(output=None):
+	"""Archives OpenBazaar files in a single tar archive."""
 	os.chdir(DATA_FOLDER)
-
-	if tableList:
-		# Parse table list
-		tableList = tableList.replace(' ', '').split(',')
-		tablesAndColumns = []
-		for table in tableList:
-			if table in TABLES:
-				tablesAndColumns.append((table, TABLES[table]))
-			else:
-				return 'ERROR, Table not found: {0}'.format(table)
-
-		# Remove existing database files and re-make them
-		if os.path.exists('backup'):
-			shutil.rmtree('backup')
-		os.makedirs('backup')
-		_exportDatabaseToCsv(tablesAndColumns)
 
 	# Archive files
 	files = os.listdir(DATA_FOLDER)
 	if not output:
 		output = 'backup_{0}.tar.gz'.format(time.strftime('%Y-%m-%d'))
+	silentRemove(output)
 	with tarfile.open(output, 'w:gz') as tar:
 		for f in files:
 			tar.add(f)
 		tar.close()
+	return 'Success'
+
+def exportDatabase(tableList, removePrevious=False):
+	"""Exports given tables to the OpenBazaar folder."""
+	# Parse table list
+	tableList = tableList.replace(' ', '').split(',')
+	tablesAndColumns = []
+	for table in tableList:
+		if table in _TABLES:
+			tablesAndColumns.append((table, _TABLES[table]))
+		else:
+			return 'ERROR, Table not found: {0}'.format(table)
+
+	# Remove existing database files and re-make them
+	if removePrevious and os.path.exists('backup'):
+		shutil.rmtree('backup')
+	if not os.path.exists('backup'):
+		os.makedirs('backup')
+	_exportDatabaseToCsv(tablesAndColumns)
 	return 'Success'
 
 def _importCsvToTable(fileName, deleteDataFirst=False):
@@ -101,27 +115,30 @@ def _importCsvToTable(fileName, deleteDataFirst=False):
 						cursor.execute(insertsql, row)
 
 
-def restore(input, deleteTableDataFirst=False):
-	"""Restores files and tables of given archive."""
+def restoreFiles(input, removePreviousDatabaseFiles=False):
+	"""Restores files of given archive to OpenBazaar folder."""
 	if not input:
 		return 'Input path is needed'
 	os.chdir(DATA_FOLDER)
 
 	# Remove existing database files if any
-	if os.path.exists('backup'):
+	if removePreviousDatabaseFiles and os.path.exists('backup'):
 		shutil.rmtree('backup')
 
 	# Unarchive files
 	with tarfile.open(input, 'r:gz') as tar:
 		tar.extractall()
 
-	# Restore database files to the database
-	if os.path.exists('backup'):
-		files = ['backup/{0}'.format(f) for f in os.listdir('backup')]
-		for f in files:
-			_importCsvToTable(f, deleteTableDataFirst)
-
 	return 'Success'
 
 if __name__ == '__main__':
 	print 'Backup tool works as a library.'
+
+def importDatabase(deletePreviousData=False):
+	"""Imports table files from the OpenBazaar folder."""
+	# Restore database files to the database
+	if os.path.exists('backup'):
+		files = ['backup/{0}'.format(f) for f in os.listdir('backup')]
+		for f in files:
+			_importCsvToTable(f, deletePreviousData)
+	return 'Success'
